Move test to its respective package
