+++ src/test/java/seedu/address/logic/commands/localcourse/LocalCourseAddCommandTest.java
@@ -1,10 +1,11 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.localcourse;
 
 import static java.util.Objects.requireNonNull;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.testutil.Assert.assertThrows;
+import static seedu.address.testutil.TypicalObjects.CS2030S;
 
 import java.nio.file.Path;
 import java.util.ArrayList;
@@ -18,9 +19,10 @@ import org.junit.jupiter.api.Test;
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.logic.Messages;
+import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.logic.commands.partnercourse.PartnerCourseAddCommand;
-import seedu.address.model.PartnerCourseCatalogue;
+import seedu.address.logic.commands.localcourse.LocalCourseAddCommand;
+import seedu.address.model.LocalCourseCatalogue;
 import seedu.address.model.ReadOnlyLocalCourseCatalogue;
 import seedu.address.model.ReadOnlyMappingCatalogue;
 import seedu.address.model.ReadOnlyNoteCatalogue;
@@ -42,65 +44,72 @@ import seedu.address.model.partnercourse.PartnerCourseContainsKeywordsPredicate;
 import seedu.address.model.university.University;
 import seedu.address.model.university.UniversityName;
 import seedu.address.model.university.UniversityNameContainsKeywordsPredicate;
-import seedu.address.testutil.PartnerCourseBuilder;
+import seedu.address.testutil.LocalCourseBuilder;
 
 /**
- * Unit testing of PartnerCourseAddCommand, with stubs / dependency injection.
+ * Unit testing of LocalCourseAddCommand, with stubs / dependency injection.
  */
-public class PartnerCourseAddCommandTest {
+public class LocalCourseAddCommandTest {
 
     @Test
-    public void constructor_nullPartnerCourse_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> new PartnerCourseAddCommand(null));
+    public void constructor_nullLocalCourse_throwsNullPointerException() {
+        assertThrows(NullPointerException.class, () -> new LocalCourseAddCommand(null));
     }
 
     @Test
-    public void execute_partnerCourseAcceptedByModel_addSuccessful() throws Exception {
-        ModelStubAcceptingPartnerCourseAdded modelStub = new ModelStubAcceptingPartnerCourseAdded();
-        PartnerCourse validPartnerCourse = new PartnerCourseBuilder().build();
+    public void execute_localCourseAcceptedByModel_addSuccessful() throws Exception {
+        ModelStubAcceptingLocalCourseAdded modelStub = new ModelStubAcceptingLocalCourseAdded();
+        LocalCourse validLocalCourse = new LocalCourseBuilder().build();
+        // Tests interactions with model
+        CommandResult commandResult = new LocalCourseAddCommand(validLocalCourse).execute(modelStub);
 
-        CommandResult commandResult = new PartnerCourseAddCommand(validPartnerCourse).execute(modelStub);
-
-        assertEquals(String.format(PartnerCourseAddCommand.MESSAGE_SUCCESS, Messages.format(validPartnerCourse)),
+        assertEquals(
+                String.format(LocalCourseAddCommand.MESSAGE_SUCCESS, Messages.format(validLocalCourse)),
                 commandResult.getFeedbackToUser());
-        assertEquals(Arrays.asList(validPartnerCourse), modelStub.partnerCoursesAdded);
+        assertEquals(Arrays.asList(validLocalCourse), modelStub.localCoursesAdded);
     }
 
     @Test
-    public void execute_duplicatePartnerCourse_throwsCommandException() {
-        PartnerCourse validPartnerCourse = new PartnerCourseBuilder().build();
-        PartnerCourseAddCommand partnerCourseAddCommand = new PartnerCourseAddCommand(validPartnerCourse);
-        SeplendidModelStub modelStub = new SeplendidModelStubWithPartnerCourse(validPartnerCourse);
+    public void execute_duplicateLocalCourse_throwsCommandException() {
+        LocalCourse validLocalCourse = new LocalCourseBuilder().build();
+        LocalCourseAddCommand localCourseAddCommand = new LocalCourseAddCommand(validLocalCourse);
+        SeplendidModelStub modelStub = new SeplendidModelStubWithLocalCourse(validLocalCourse);
 
         assertThrows(
                 CommandException.class,
-                PartnerCourseAddCommand.MESSAGE_DUPLICATE_PARTNER_COURSE, (
-                ) -> partnerCourseAddCommand.execute(modelStub));
+                LocalCourseAddCommand.MESSAGE_DUPLICATE_LOCAL_COURSE, (
+                ) -> localCourseAddCommand.execute(modelStub));
     }
 
     @Test
     public void equals() {
-        PartnerCourse cs1234 = new PartnerCourseBuilder().withPartnerCode("CS1234").build();
-        PartnerCourse cs6789 = new PartnerCourseBuilder().withPartnerCode("CS6789").build();
-        PartnerCourseAddCommand addCS1234Command = new PartnerCourseAddCommand(cs1234);
-        PartnerCourseAddCommand addCS6789Command = new PartnerCourseAddCommand(cs6789);
+        LocalCourse cs2030s = new LocalCourseBuilder().withLocalCode("CS2030S").build();
+        LocalCourse cs2040s = new LocalCourseBuilder().withLocalCode("CS2040S").build();
+        LocalCourseAddCommand addCS2030SCommand = new LocalCourseAddCommand(cs2030s);
+        LocalCourseAddCommand addCS2040SCommand = new LocalCourseAddCommand(cs2040s);
 
         // same object -> returns true
-        assertTrue(addCS1234Command.equals(addCS1234Command));
+        assertTrue(addCS2030SCommand.equals(addCS2030SCommand));
 
         // same values -> returns true
-        PartnerCourseAddCommand addCS1234CommandCopy = new PartnerCourseAddCommand(cs1234);
-        assertTrue(addCS1234Command.equals(addCS1234CommandCopy));
+        LocalCourseAddCommand addCS2030SCommandCopy = new LocalCourseAddCommand(cs2030s);
+        assertTrue(addCS2030SCommand.equals(addCS2030SCommandCopy));
 
         // different types -> returns false
-        assertFalse(addCS1234Command.equals(1));
+        assertFalse(addCS2030SCommand.equals(1));
 
         // null -> returns false
-        assertFalse(addCS1234Command.equals(null));
+        assertFalse(addCS2030SCommand.equals(null));
 
         // different local course -> returns false
-        assertFalse(addCS1234Command.equals(addCS6789Command));
+        assertFalse(addCS2030SCommand.equals(addCS2040SCommand));
+    }
 
+    @Test
+    public void toStringMethod() {
+        LocalCourseAddCommand localCourseAddCommand = new LocalCourseAddCommand(CS2030S);
+        String expected = LocalCourseAddCommand.class.getCanonicalName() + "{localCourseToAdd=" + CS2030S + "}";
+        assertEquals(expected, localCourseAddCommand.toString());
     }
 
     /**
@@ -211,12 +220,12 @@ public class PartnerCourseAddCommandTest {
         }
 
         @Override
-        public void setPartnerCourseCatalogueFilePath(Path partnerCourseCatalogueFilePath) {
+        public Path getPartnerCourseCatalogueFilePath() {
             throw new AssertionError("This method should not be called.");
         }
 
         @Override
-        public Path getPartnerCourseCatalogueFilePath() {
+        public void setPartnerCourseCatalogueFilePath(Path partnerCourseCatalogueFilePath) {
             throw new AssertionError("This method should not be called.");
         }
 
@@ -288,22 +297,12 @@ public class PartnerCourseAddCommandTest {
         }
 
         @Override
-        public void setUniversityCatalogueFilePath(Path universityCatalogueFilePath) {
-        }
-
         public Optional<University> getUniversityIfExists(UniversityName universityName) {
             throw new AssertionError("This method should not be called.");
         }
 
         @Override
         public void setNoteCatalogue(ReadOnlyNoteCatalogue noteCatalogue) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void getSearchUniversityIfExists(
-                UniversityNameContainsKeywordsPredicate universityNameContainsKeywordsPredicate
-        ) {
             throw new AssertionError("This method should not be called.");
         }
 
@@ -319,6 +318,11 @@ public class PartnerCourseAddCommandTest {
 
         @Override
         public Note deleteNote(int noteIndex) {
+            throw new AssertionError("This method should not be called.");
+        }
+
+        @Override
+        public void setUniversityCatalogueFilePath(Path universityCatalogueFilePath) {
             throw new AssertionError("This method should not be called.");
         }
 
@@ -343,6 +347,13 @@ public class PartnerCourseAddCommandTest {
         }
 
         @Override
+        public void getSearchUniversityIfExists(
+                UniversityNameContainsKeywordsPredicate universityNameContainsKeywordsPredicate
+        ) {
+            throw new AssertionError("This method should not be called.");
+        }
+
+        @Override
         public ObservableList<University> getFilteredUniversityList() {
             throw new AssertionError("This method should not be called.");
         }
@@ -351,6 +362,7 @@ public class PartnerCourseAddCommandTest {
         public boolean hasUniversity(University university) {
             throw new AssertionError("This method should not be called.");
         }
+
 
         @Override
         public boolean hasUniversity(UniversityName universityName) {
@@ -468,60 +480,48 @@ public class PartnerCourseAddCommandTest {
         public void updateSortedMappingList(Comparator<Mapping> mappingComparator) {
             throw new AssertionError("This method should not be called.");
         }
-
     }
 
     /**
-     * A Model stub that contains a single partnerCourse.
+     * A Model stub that contains a single localCourse.
      */
-    private class SeplendidModelStubWithPartnerCourse extends PartnerCourseAddCommandTest.SeplendidModelStub {
-        private final PartnerCourse partnerCourse;
+    private class SeplendidModelStubWithLocalCourse extends SeplendidModelStub {
+        private final LocalCourse localCourse;
 
-        SeplendidModelStubWithPartnerCourse(PartnerCourse pc) {
-            requireNonNull(pc);
-            partnerCourse = pc;
+        SeplendidModelStubWithLocalCourse(LocalCourse lc) {
+            requireNonNull(lc);
+            localCourse = lc;
         }
 
         @Override
-        public boolean hasPartnerCourse(PartnerCourse pc) {
-            requireNonNull(pc);
-            return partnerCourse.isSamePartnerCourse(pc);
-        }
-
-        @Override
-        public boolean hasUniversity(University university) {
-            requireNonNull(university);
-            return partnerCourse.getPartnerUniversity().isSameUniversity(university);
+        public boolean hasLocalCourse(LocalCourse lc) {
+            requireNonNull(lc);
+            return localCourse.isSameLocalCourse(lc);
         }
     }
 
-    private class ModelStubAcceptingPartnerCourseAdded extends PartnerCourseAddCommandTest.SeplendidModelStub {
-        final ArrayList<PartnerCourse> partnerCoursesAdded = new ArrayList<>();
+    /**
+     * A Model stub that always accept the local course being added.
+     */
+    private class ModelStubAcceptingLocalCourseAdded extends SeplendidModelStub {
+        final ArrayList<LocalCourse> localCoursesAdded = new ArrayList<>();
 
         @Override
-        public boolean hasPartnerCourse(PartnerCourse partnerCourse) {
-            requireNonNull(partnerCourse);
-            return partnerCoursesAdded.stream().anyMatch(partnerCourse::isSamePartnerCourse);
+        public boolean hasLocalCourse(LocalCourse localCourse) {
+            requireNonNull(localCourse);
+            return localCoursesAdded.stream().anyMatch(localCourse::isSameLocalCourse);
         }
 
         @Override
-        public void addPartnerCourse(PartnerCourse partnerCourse) {
-            requireNonNull(partnerCourse);
-            partnerCoursesAdded.add(partnerCourse);
+        public void addLocalCourse(LocalCourse localCourse) {
+            requireNonNull(localCourse);
+            localCoursesAdded.add(localCourse);
         }
 
         @Override
-        public boolean hasUniversity(University university) {
-            requireNonNull(university);
-            //assume that the university exists
-            return true;
-        }
-
-        @Override
-        public ReadOnlyPartnerCourseCatalogue getPartnerCourseCatalogue() {
-            return new PartnerCourseCatalogue();
+        public ReadOnlyLocalCourseCatalogue getLocalCourseCatalogue() {
+            return new LocalCourseCatalogue();
         }
     }
-
 
 }
