+++ src/test/java/seedu/address/logic/commands/localcourse/LocalCourseListCommandTest.java
@@ -1,7 +1,7 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.localcourse;
 
 import static seedu.address.logic.commands.CommandTestUtil.assertSeplendidCommandSuccess;
-import static seedu.address.logic.commands.CommandTestUtil.showPartnerCourseAtIndex;
+import static seedu.address.logic.commands.CommandTestUtil.showLocalCourseAtIndex;
 import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_OBJECT;
 import static seedu.address.testutil.TypicalObjects.getTypicalLocalCourseCatalogue;
 import static seedu.address.testutil.TypicalObjects.getTypicalMappingCatalogue;
@@ -12,7 +12,7 @@ import static seedu.address.testutil.TypicalObjects.getTypicalUniversityCatalogu
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.partnercourse.PartnerCourseListCommand;
+import seedu.address.logic.commands.localcourse.LocalCourseListCommand;
 import seedu.address.model.SeplendidModel;
 import seedu.address.model.SeplendidModelManager;
 import seedu.address.model.UserPrefs;
@@ -20,7 +20,8 @@ import seedu.address.model.UserPrefs;
 /**
  * Contains integration tests (interaction with the Model) and unit tests for ListCommand.
  */
-public class PartnerCourseListCommandTest {
+public class LocalCourseListCommandTest {
+
     private SeplendidModel model;
     private SeplendidModel expectedModel;
 
@@ -33,27 +34,27 @@ public class PartnerCourseListCommandTest {
                                           getTypicalMappingCatalogue(),
                                           getTypicalNoteCatalogue());
         expectedModel = new SeplendidModelManager(new UserPrefs(),
-                                                  getTypicalLocalCourseCatalogue(),
-                                                  model.getPartnerCourseCatalogue(),
+                                                  model.getLocalCourseCatalogue(),
+                                                  getTypicalPartnerCourseCatalogue(),
                                                   getTypicalUniversityCatalogue(),
                                                   getTypicalMappingCatalogue(),
                                                   getTypicalNoteCatalogue());
     }
 
     @Test
-    public void execute_listNotFiltered_showsSameList() {
-        assertSeplendidCommandSuccess(new PartnerCourseListCommand(),
+    public void execute_listIsNotFiltered_showsSameList() {
+        assertSeplendidCommandSuccess(new LocalCourseListCommand(),
                                       model,
-                                      PartnerCourseListCommand.MESSAGE_SUCCESS,
+                                      LocalCourseListCommand.MESSAGE_SUCCESS,
                                       expectedModel);
     }
 
     @Test
     public void execute_listIsFiltered_showsEverything() {
-        showPartnerCourseAtIndex(model, INDEX_FIRST_OBJECT);
-        assertSeplendidCommandSuccess(new PartnerCourseListCommand(),
+        showLocalCourseAtIndex(model, INDEX_FIRST_OBJECT);
+        assertSeplendidCommandSuccess(new LocalCourseListCommand(),
                                       model,
-                                      PartnerCourseListCommand.MESSAGE_SUCCESS,
+                                      LocalCourseListCommand.MESSAGE_SUCCESS,
                                       expectedModel);
     }
 }
