+++ src/test/java/seedu/address/logic/commands/partnercourse/PartnerCourseListCommandTest.java
@@ -1,7 +1,7 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.partnercourse;
 
 import static seedu.address.logic.commands.CommandTestUtil.assertSeplendidCommandSuccess;
-import static seedu.address.logic.commands.CommandTestUtil.showUniversityAtIndex;
+import static seedu.address.logic.commands.CommandTestUtil.showPartnerCourseAtIndex;
 import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_OBJECT;
 import static seedu.address.testutil.TypicalObjects.getTypicalLocalCourseCatalogue;
 import static seedu.address.testutil.TypicalObjects.getTypicalMappingCatalogue;
@@ -12,12 +12,15 @@ import static seedu.address.testutil.TypicalObjects.getTypicalUniversityCatalogu
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.university.UniversityListCommand;
+import seedu.address.logic.commands.partnercourse.PartnerCourseListCommand;
 import seedu.address.model.SeplendidModel;
 import seedu.address.model.SeplendidModelManager;
 import seedu.address.model.UserPrefs;
 
-public class UniversityListCommandTest {
+/**
+ * Contains integration tests (interaction with the Model) and unit tests for ListCommand.
+ */
+public class PartnerCourseListCommandTest {
     private SeplendidModel model;
     private SeplendidModel expectedModel;
 
@@ -31,27 +34,26 @@ public class UniversityListCommandTest {
                                           getTypicalNoteCatalogue());
         expectedModel = new SeplendidModelManager(new UserPrefs(),
                                                   getTypicalLocalCourseCatalogue(),
-                                                  getTypicalPartnerCourseCatalogue(),
-                                                  model.getUniversityCatalogue(),
+                                                  model.getPartnerCourseCatalogue(),
+                                                  getTypicalUniversityCatalogue(),
                                                   getTypicalMappingCatalogue(),
                                                   getTypicalNoteCatalogue());
     }
 
     @Test
-    public void execute_listIsNotFiltered_showsSameList() {
-        assertSeplendidCommandSuccess(new UniversityListCommand(),
+    public void execute_listNotFiltered_showsSameList() {
+        assertSeplendidCommandSuccess(new PartnerCourseListCommand(),
                                       model,
-                                      UniversityListCommand.MESSAGE_SUCCESS,
+                                      PartnerCourseListCommand.MESSAGE_SUCCESS,
                                       expectedModel);
     }
 
     @Test
     public void execute_listIsFiltered_showsEverything() {
-        showUniversityAtIndex(model, INDEX_FIRST_OBJECT);
-        assertSeplendidCommandSuccess(new UniversityListCommand(),
+        showPartnerCourseAtIndex(model, INDEX_FIRST_OBJECT);
+        assertSeplendidCommandSuccess(new PartnerCourseListCommand(),
                                       model,
-                                      UniversityListCommand.MESSAGE_SUCCESS,
+                                      PartnerCourseListCommand.MESSAGE_SUCCESS,
                                       expectedModel);
-
     }
 }
