+++ src/test/java/seedu/address/logic/commands/university/UniversityListCommandTest.java
@@ -1,7 +1,7 @@
-package seedu.address.logic.commands;
+package seedu.address.logic.commands.university;
 
 import static seedu.address.logic.commands.CommandTestUtil.assertSeplendidCommandSuccess;
-import static seedu.address.logic.commands.CommandTestUtil.showLocalCourseAtIndex;
+import static seedu.address.logic.commands.CommandTestUtil.showUniversityAtIndex;
 import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_OBJECT;
 import static seedu.address.testutil.TypicalObjects.getTypicalLocalCourseCatalogue;
 import static seedu.address.testutil.TypicalObjects.getTypicalMappingCatalogue;
@@ -12,16 +12,12 @@ import static seedu.address.testutil.TypicalObjects.getTypicalUniversityCatalogu
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import seedu.address.logic.commands.localcourse.LocalCourseListCommand;
+import seedu.address.logic.commands.university.UniversityListCommand;
 import seedu.address.model.SeplendidModel;
 import seedu.address.model.SeplendidModelManager;
 import seedu.address.model.UserPrefs;
 
-/**
- * Contains integration tests (interaction with the Model) and unit tests for ListCommand.
- */
-public class LocalCourseListCommandTest {
-
+public class UniversityListCommandTest {
     private SeplendidModel model;
     private SeplendidModel expectedModel;
 
@@ -34,27 +30,28 @@ public class LocalCourseListCommandTest {
                                           getTypicalMappingCatalogue(),
                                           getTypicalNoteCatalogue());
         expectedModel = new SeplendidModelManager(new UserPrefs(),
-                                                  model.getLocalCourseCatalogue(),
+                                                  getTypicalLocalCourseCatalogue(),
                                                   getTypicalPartnerCourseCatalogue(),
-                                                  getTypicalUniversityCatalogue(),
+                                                  model.getUniversityCatalogue(),
                                                   getTypicalMappingCatalogue(),
                                                   getTypicalNoteCatalogue());
     }
 
     @Test
     public void execute_listIsNotFiltered_showsSameList() {
-        assertSeplendidCommandSuccess(new LocalCourseListCommand(),
+        assertSeplendidCommandSuccess(new UniversityListCommand(),
                                       model,
-                                      LocalCourseListCommand.MESSAGE_SUCCESS,
+                                      UniversityListCommand.MESSAGE_SUCCESS,
                                       expectedModel);
     }
 
     @Test
     public void execute_listIsFiltered_showsEverything() {
-        showLocalCourseAtIndex(model, INDEX_FIRST_OBJECT);
-        assertSeplendidCommandSuccess(new LocalCourseListCommand(),
+        showUniversityAtIndex(model, INDEX_FIRST_OBJECT);
+        assertSeplendidCommandSuccess(new UniversityListCommand(),
                                       model,
-                                      LocalCourseListCommand.MESSAGE_SUCCESS,
+                                      UniversityListCommand.MESSAGE_SUCCESS,
                                       expectedModel);
+
     }
 }
